refactor: clean up repo infrastructure and file organization
