+++ .gitignore
@@ -1,5 +1,7 @@
 # custom
 dist
+build
+package
 examples/**/package-lock.json
 .svelte-kit
