--- decs.d.ts
@@ -1,3 +1,0 @@
-declare module "sm-commons";
-declare module "rollup-plugin-includepaths";
-declare module "sm-commons/utils/str";
