+++ package.json
@@ -15,54 +15,52 @@
 	"author": "Prismic <contact@prismic.io> (https://prismic.io)",
 	"type": "module",
 	"exports": {
-		"./package.json": "./package.json",
 		".": {
 			"types": "./dist/index.d.ts",
-			"svelte": "./dist/index.js",
-			"default": "./dist/index.js"
+			"svelte": "./dist/index.js"
 		},
-		"./index.d.ts": "./dist/index.d.ts"
+		"./package.json": "./package.json"
 	},
-	"main": "./dist/index.js",
-	"module": "./dist/index.js",
-	"types": "./dist/index.d.ts",
 	"svelte": "./dist/index.js",
+	"types": "./dist/index.d.ts",
+	"files": [
+		"dist",
+		"src"
+	],
 	"scripts": {
-		"build": "svelte-package -o dist package && shx cp -r src dist/src",
+		"build": "vite build && svelte-kit sync && svelte-package",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"dev": "vite dev",
 		"format": "prettier --write .",
+		"lint": "eslint --ext .js,.ts .",
 		"prepare": "npm run build",
+		"preview": "vite preview",
 		"release": "npm run build && npm run test && standard-version && git push --follow-tags && npm run build && npm publish",
-		"release:dry": "standard-version --dry-run",
 		"release:alpha": "npm run build && npm run test && standard-version --release-as major --prerelease alpha && git push --follow-tags && npm run build && npm publish --tag alpha",
 		"release:alpha:dry": "standard-version --release-as major --prerelease alpha --dry-run",
 		"release:dev": "npm run build && npm run test && standard-version --prerelease dev && git push --follow-tags && npm run build && npm publish --tag dev",
 		"release:dev:dry": "standard-version --prerelease dev --dry-run",
-		"lint": "eslint --ext .js,.ts .",
-		"unit": "vitest run --coverage",
-		"unit:watch": "vitest watch",
+		"release:dry": "standard-version --dry-run",
 		"test": "npm run lint && npm run unit",
-		"preview": "svelte-kit preview",
-		"package": "svelte-package -o dist package",
-		"check": "svelte-check --tsconfig ./tsconfig.json",
-		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch"
+		"unit": "vitest run --coverage",
+		"unit:watch": "vitest watch"
 	},
 	"dependencies": {
 		"@prismicio/client": "^6.7.3",
 		"@prismicio/helpers": "^2.3.9",
 		"@prismicio/richtext": "^2.2.0-alpha.0",
-		"@sveltejs/kit": "^1.8.3",
-		"svelte": "^3.55.1"
+		"esm-env": "^1.0.0"
 	},
 	"devDependencies": {
 		"@prismicio/mock": "^0.2.0",
 		"@prismicio/types": "^0.2.7",
-		"@rollup/plugin-inject": "^5.0.3",
+		"@sveltejs/adapter-auto": "^2.0.0",
+		"@sveltejs/kit": "^1.8.3",
 		"@sveltejs/package": "^2.0.0",
 		"@typescript-eslint/eslint-plugin": "^5.53.0",
 		"@typescript-eslint/parser": "^5.53.0",
 		"@vitest/coverage-c8": "^0.28.5",
-		"c8": "^7.13.0",
 		"eslint": "^8.15.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-prettier": "^4.2.0",
@@ -71,77 +69,20 @@
 		"prettier": "^2.8.4",
 		"prettier-plugin-jsdoc": "^0.3.38",
 		"prettier-plugin-svelte": "^2.9.0",
-		"shx": "^0.3.4",
 		"standard-version": "^9.3.2",
+		"svelte": "^3.55.1",
 		"svelte-check": "^3.0.3",
-		"svelte-preprocess": "^5.0.1",
-		"svelte2tsx": "^0.6.1",
 		"tslib": "^2.5.0",
 		"typescript": "^4.9.5",
 		"vitest": "^0.28.5"
+	},
+	"peerDependencies": {
+		"svelte": "^3.54.0"
 	},
 	"engines": {
 		"node": ">=12.7.0"
 	},
 	"publishConfig": {
 		"access": "public"
-	},
-	"optionalDependencies": {
-		"@esbuild/android-arm": "^0.17.10"
-	},
-	"files": [
-		"dist"
-	],
-	"typesVersions": {
-		">4.0": {
-			"PrismicProvider.svelte": [
-				"./dist/PrismicProvider.svelte.d.ts"
-			],
-			"components/PrismicEmbed.svelte": [
-				"./dist/components/PrismicEmbed.svelte.d.ts"
-			],
-			"components/PrismicImage.svelte": [
-				"./dist/components/PrismicImage.svelte.d.ts"
-			],
-			"components/PrismicLink.svelte": [
-				"./dist/components/PrismicLink.svelte.d.ts"
-			],
-			"components/PrismicPreview.svelte": [
-				"./dist/components/PrismicPreview.svelte.d.ts"
-			],
-			"components/PrismicRichText/DefaultComponent.svelte": [
-				"./dist/components/PrismicRichText/DefaultComponent.svelte.d.ts"
-			],
-			"components/PrismicRichText/PrismicRichText.svelte": [
-				"./dist/components/PrismicRichText/PrismicRichText.svelte.d.ts"
-			],
-			"components/PrismicRichText/Serialize.svelte": [
-				"./dist/components/PrismicRichText/Serialize.svelte.d.ts"
-			],
-			"components/PrismicText.svelte": [
-				"./dist/components/PrismicText.svelte.d.ts"
-			],
-			"components/SliceZone.svelte": [
-				"./dist/components/SliceZone.svelte.d.ts"
-			],
-			"components/TodoSliceComponent.svelte": [
-				"./dist/components/TodoSliceComponent.svelte.d.ts"
-			],
-			"components": [
-				"./dist/components/index.d.ts"
-			],
-			"configurePrismic": [
-				"./dist/configurePrismic.d.ts"
-			],
-			"index": [
-				"./dist/index.d.ts"
-			],
-			"injectionSymbols": [
-				"./dist/injectionSymbols.d.ts"
-			],
-			"usePrismic": [
-				"./dist/usePrismic.d.ts"
-			]
-		}
 	}
 }
