+++ src/lib/SliceZone/SliceZone.svelte
@@ -80,18 +80,12 @@
 		| SvelteComponent
 		| (new (...args: any[]) => SvelteComponent)
 		| undefined = undefined;
-
-	/**
-	 * Whether the app is in development mode. When true, the SliceZone may
-	 * display error messages in the UI.
-	 */
-	export let dev: boolean = false;
 </script>
 
-<!-- 
+<!--
   @component
   Component to render an array of Prismic Slices.
-  
+
   @example Rendering a Rich Text field:
 	```svelte
 		<SliceZone slices={document.data.slices} components={{
@@ -107,6 +101,6 @@
 	{#if Component}
 		<svelte:component this={Component} {slice} {slices} {context} {index} />
 	{:else}
-		<TodoSliceComponent {slice} {dev} />
+		<TodoSliceComponent {slice} />
 	{/if}
 {/each}
