+++ src/lib/index.ts
@@ -1,9 +1,6 @@
-export {
-	// Components
-	PrismicEmbed,
-	PrismicImage,
-	PrismicLink,
-	PrismicText,
-	PrismicRichText,
-	SliceZone,
-} from "./components";
+export { default as PrismicEmbed } from "./PrismicEmbed.svelte";
+export { default as PrismicImage } from "./PrismicImage.svelte";
+export { default as PrismicLink } from "./PrismicLink.svelte";
+export { default as PrismicRichText } from "./PrismicRichText/PrismicRichText.svelte";
+export { default as PrismicText } from "./PrismicText.svelte";
+export { default as SliceZone } from "./SliceZone/SliceZone.svelte";
