--- src/routes/components/index.js
@@ -1,1 +1,0 @@
-export { default as Em } from "./Emphasis.svelte";
