+++ src/routes/slices/CodeSnippet.svelte
