--- src/routes/slices/index.ts
@@ -1,2 +1,0 @@
-export { default as CodeSnippet } from "./CodeSnippet/index.svelte";
-export { default as RichText } from "./RichText/index.svelte";
