+++ svelte.config.js
@@ -1,12 +1,20 @@
-/* eslint-disable tsdoc/syntax */
-import preprocess from "svelte-preprocess";
+import adapter from "@sveltejs/adapter-auto";
+import { vitePreprocess } from "@sveltejs/kit/vite";
 
 /**
  * @type {import("@sveltejs/kit").Config}
  */
 const config = {
-	preprocess: preprocess(),
-	kit: {},
+	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
+	// for more information about preprocessors
+	preprocess: vitePreprocess(),
+
+	kit: {
+		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
+		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
+		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
+		adapter: adapter(),
+	},
 };
 
 export default config;
