--- vitest.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-	test: {
-		coverage: {
-			reporter: ["lcovonly", "text"],
-		},
-	},
-});
